drf-cookiecutter/{{ cookiecutter.main_folder }}/requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
dj-database-url==1.2.0
djangorestframework==3.14.0
djangorestframework_simplejwt[crypto]==5.2.2
gunicorn==20.1.0
psycopg2-binary==2.9.6
sentry-sdk==1.13.0
requests==2.31.0
